drf/requirements.txt
Django==3.2.13
mysqlclient==2.0.3
djangorestframework==3.13.1
django-filter==21.1
django-allauth==0.50.0
djangorestframework-simplejwt==5.2.0